==> marine/shrimp/shrimp_farms/STEP2_shrimp_farm_locations.Rmd <==
---
title: "Location of Shrimp Farms"
author: "Gage Clawson (USCB, NCEAS, OHI)"
date: "10/28/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

This script standardizes location of shrimp farms to be lat/long point data.


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(raster)
library(tidyverse)
library(stringr)
library(rgdal)
library(PBSmapping)
library(maptools)
library(sf)
library(lwgeom)
library(data.table)
library(here)
library(mapview)
library(janitor)
library(rnaturalearth)
library(spatialEco)
library(countrycode)

path <- "/home/shares/food-systems/Food_footprint/_raw_data/Aquaculture_locations"
fao_shrimp <- read_csv("marine/shrimp/shrimp_farms/data/fao_mariculture_shrimp.csv") %>%
  dplyr::filter(year == 2017) 


unique(fao_shrimp$country)
#  [1] "Australia"                "Bangladesh"               "Belize"                   "Brazil"                  
#  [5] "Brunei Darussalam"        "Cambodia"                 "China"                    "Colombia"                
#  [9] "Costa Rica"               "Cuba"                     "Ecuador"                  "El Salvador"             
# [13] "Guatemala"                "Honduras"                 "India"                    "Indonesia"               
# [17] "Iran (Islamic Rep. of)"   "Japan"                    "Korea, Republic of"       "Madagascar"              
# [21] "Malaysia"                 "Mexico"                   "Myanmar"                  "New Caledonia"           
# [25] "Nicaragua"                "Panama"                   "Peru"                     "Philippines"             
# [29] "Saudi Arabia"             "Sri Lanka"                "Taiwan Province of China" "Thailand"                
# [33] "United States of America" "Venezuela, Boliv Rep of"  "Viet Nam" 

```


# Raster

Standard raster used for this analysis 

```{r}

source(here::here("_spatial/template_raster.R"))

```

# Global regions

```{r}

gadm <- raster("/home/shares/food-systems/Food_footprint/_raw_data/Country_regions/gadm36mol_pyfill_10km.tif")

gadm_latlon <- projectRaster(gadm, food_raster)

gadm_codes <- read.csv("/home/shares/food-systems/Food_footprint/_raw_data/Country_regions/gadm_reg_codes.csv") %>% 
  mutate(reg_id = as.numeric(reg_id))

```

```{r}

gadm_df <- tabularaster::as_tibble(gadm_latlon, value = TRUE, cells = TRUE) %>% 
  mutate(cellvalue = round(cellvalue, 0))

gadm_reg <- dplyr::left_join(gadm_df, gadm_codes, by = c("cellvalue" = "reg_id"))

```

Now we will map the remaining countries with shrimp farms. Workflow will be as follows: 
 - Try to obtain exact locations of farms from published source or 
 dataset (such as FAO or a research paper)
 - If there are not exact locations or comprehensive locations, find the number of farms. 
 - If there is no information at all, note it. We will gapfill later. 
 
For each point, we will add a column for "data quality". It will be scored as very good (5), good (4), fair (3), bad (2), and very bad (1). 

5: Comprehensive number of exact farms 

4: Subset of exact locations, need to add a specific number of farms to existing locations. 
  - country|species|farm locations
  - country|species|additional points

3: A specific number of farms are known, but no locations 
 - country|species|additional points
 
2.5: A specific number of farms that are not comprehensive, so we will have to gapfill on top of those, if necessary. 
 - country, species, n_existing_points

2: Exact locations are a subset of total farms. But the number of total farms are unknown. Need to be gapfilled by tonnage. 

1: No information on number of farms at all. Need to be gapfilled by tonnage. 
 - country|species|# additional farms (this will be NA)

## Read in ASC data
https://gis.asc-aqua.org/arcgis_app/
Maps from ASC

```{r}
ASC_data <- st_read(file.path(path, "ASC_data/ASC_all_sf.shp"))

unique(ASC_data$spcs__c)

ASC_shrimp <- ASC_data %>%
  dplyr::filter(spcs__c == "Shrimp")

countries <- c(sort(unique(ASC_shrimp$crt_cnt)))

#  [1] "India" (FAO)                              "Viet Nam"    (FAO)                      
#  [3] "Ecuador"    (FAO)                        "Venezuela (Bolivarian Republic of)" (FAO)
#  [5] "Honduras"      (FAO)                     "China"           (FAO)                  
#  [7] "Mexico"      (FAO)                       "Thailand"                          
#  [9] "Madagascar"    (FAO)                     "Panama"   (FAO)                         
# [11] "Indonesia"  (FAO)                        "Nigeria"    (NO FAO)                       
# [13] "Bangladesh" (FAO)                        "Guatemala"    (FAO)                     
# [15] "Nicaragua"     (FAO)                     "Belize"         (FAO)                   
# [17] "New Caledonia"  (FAO)                    "Tanzania, United Republic of"      (NO FAO)
# [19] "Australia"      (FAO)                    "Mozambique"  (NO FAO)                      
# [21] "Peru"    (FAO)                           "Colombia"   (FAO)


setdiff(ASC_shrimp$crt_cnt, fao_shrimp$country) # don't worry about venezuela... just a mispelling mismatch
# In ASC and not in FAO:
# [1] "Nigeria"                           
# [3] "Tanzania, United Republic of"       "Mozambique"  

setdiff(fao_shrimp$country, ASC_shrimp$crt_cnt)
## Need to find the farm locations for these countries: 

#  [1] "Brazil"                   "Brunei Darussalam"        "Cambodia"                 "Costa Rica"              
#  [5] "Cuba"                     "El Salvador"              "Iran (Islamic Rep. of)"   "Japan"                   
#  [9] "Korea, Republic of"       "Malaysia"                 "Myanmar"                  "Philippines"             
# [13] "Saudi Arabia"             "Sri Lanka"                "Taiwan Province of China" "United States of America"

```


# Map data from ASC (contains 19/35 countries)

Data comes from https://gis.asc-aqua.org/arcgis_app/ 

Source: 
Aquaculture Stewardship Council Data Mapping Study, n.d. Aquaculture Stewardship Council. [Accessed November 11, 2020] 
https://gis.asc-aqua.org/arcgis_app/


```{r}

# get iso3c codes
iso <-   data.frame(iso3c = countrycode(ASC_shrimp$crt_cnt, origin = 'country.name', destination = 'iso3c'))


# read in shp and filter out countries not in FAO
farms_all <- ASC_shrimp %>%
  cbind(iso) %>%
  dplyr::filter(iso3c != "NGA",
                iso3c != "TZA",
                iso3c != "MOZ")

mapview(farms_all$geometry)

unique(farms_all$iso3c)

all_pts <- st_centroid(farms_all)

all_pts_latlon <- st_transform(all_pts, crs = "+init=epsg:4326") %>% 
  dplyr::select(geometry, iso3c)

farms_all_fin <- data.frame(do.call(rbind, st_geometry(all_pts_latlon))) 
names(farms_all_fin) <- c("lon", "lat")
farms_all_fin <- farms_all_fin %>%
  cbind(iso3c = all_pts_latlon$iso3c) %>%
  mutate(data_quality = 5)

## save farm location
write.csv(farms_all_fin, here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv"), row.names = FALSE)

```

Now we need to map for each country, and assign a flag for gapfilling which will be completed later. I will start with all of the countries list in the ASC data. 

# Indonesia 
Current data is from ASC, only 5 farms, so probably not comprehensive, just a drop in the bucket compared to the production of Indonesia. There are likely many more farms than this. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

```{r}
farms_idn <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "IDN") %>%
  mutate(data_quality = 2)

write_csv(farms_idn, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_idn.csv"))

```


# New Caledonia 

Current data is from ASC, only 2 farms, so probably not comprehensive. There are likely many more farms than this. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

```{r}
farms_ncl <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "NCL") %>%
  mutate(data_quality = 2)

write_csv(farms_ncl, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_ncl.csv"))


```

# Mexico 
Current data is from ASC, only 11 farms, so probably not comprehensive, just a drop in the bucket compared to the produciton of Mexico. There are likely many more farms than this. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

```{r}
farms_mex <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "MEX") %>%
  mutate(data_quality = 2)

write_csv(farms_mex, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_mex.csv"))


```

# Venezuela 
Current data is from ASC, only 11 farms, so probably not comprehensive, just a drop in the bucket compared to the produciton of Venezuela. There are likely many more farms than this. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

```{r}
farms_ven <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "VEN") %>%
  mutate(data_quality = 2)

write_csv(farms_ven, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_ven.csv"))


```



# India 

Current data is from ASC, only 104 farms, so probably not comprehensive, just a drop in the bucket compared to the produciton of India. There are likely many more farms than this. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

```{r}
farms_ind <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "IND") %>%
  mutate(data_quality = 2)

write_csv(farms_ind, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_ind.csv"))


```

# Viet Nam

Current data is from ASC, only 173 farms, so probably not comprehensive, just a drop in the bucket compared to the produciton of viet nam. There are likely many more farms than this. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

```{r}
farms_vnm <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "VNM") %>%
  mutate(data_quality = 2)

write_csv(farms_vnm, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_vnm.csv"))


```

# Ecuador

Current data is from ASC, which is likely not comprehensive. 

The FAO states this: 
"Until 1998 (the last period for which statistics on the subject area available), the Under-Ministry of Fisheries Resources registered 2 006 shrimp industries, 312 larvae hatcheries, 21 feed manufacturing plants and 76 processing plants."

It is hard to tell how many shrimp farms are actually in Ecuador. I do think the ASC number is a good number, but it is probably not comprehensive given the amount of production from Ecuador. Data quality = 2. 


```{r}
farms_ecu <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "ECU") %>%
  mutate(data_quality = 2)

write_csv(farms_ecu, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_ecu.csv"))


```


# Honduras

Current data is from ASC, only 20 farms, so probably not comprehensive. There are likely many more farms than this. The FAO states there are many smal lscale shrimp farms tat are not documented, so it is impossible to tell how many there are from current information. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

```{r}
farms_hnd <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "HND") %>%
  mutate(data_quality = 2)

write_csv(farms_hnd, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_hnd.csv"))


```

# China

Current data is from https://zenodo.org/record/3881612 and ASC, only 435 + 4 = 439 farms, so probably not comprehensive, just a drop in the bucket compared to the produciton of China. There are likely many more farms than this. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

```{r}
farms_chn <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "CHN") %>%
  mutate(data_quality = 2)


farms_chn_2 <- read.csv("/home/shares/food-systems/Food_footprint/_raw_data/Aquaculture_locations/China/split_spp/farms_shrimp_chn.csv") %>%
  dplyr::select(-id) %>%
  mutate(data_quality = 2)

farms_chn_fin <- rbind(farms_chn, farms_chn_2)

write_csv(farms_chn_fin, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_chn.csv"))


```

# Thailand
Current data is from ASC, only 18 farms, so definitely not comprehensive, just a drop in the bucket compared to the produciton of Thailand. There are likely many more farms than this. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

```{r}
farms_tha <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "THA") %>%
  mutate(data_quality = 2)

write_csv(farms_tha, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_tha.csv"))


```

# Madagasgar

Need to add 4 points to existing 2 (a final total of 6 points)

From FAO: http://www.fao.org/fishery/countrysector/naso_madagascar/en

"For marine aquaculture, the **six shrimp breeding companies** created 4 325 full-time jobs, 4 267 of which were for local people and 58 for expatriates (Anon. 2005b). These 4 325 full-time jobs were distributed as follows: 90 mariners, 1 072 in the processing plant, 400 in administration, 267 in hatcheries, 2 114 in farms, and 382 in other areas. This kind of shrimp farming also created 30 000 indirect jobs. Seaweed culture was being practised by 126 farmers."

Data quality = 4


```{r}
farms_mdg <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "MDG") %>%
  mutate(data_quality = 4)

write_csv(farms_mdg, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_mdg.csv"))

## Now we will make a file that tells us the number of additional farms that need to be added to be comprehensive
mdg_additional <- data.frame(iso3c = "MDG", type = "shrimp", num_additional_farms = 4, data_quality = 4, total_farms = 6)
write.csv(mdg_additional, here("marine/shrimp/shrimp_farms/data/subset_known_add_points_4/add_farms_mdg.csv"), row.names = FALSE)
```

# Panama
Current data is from ASC, only 1 farms, so definitely not comprehensive, just a drop in the bucket compared to the produciton of Panama. There are likely more farms than this. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

```{r}
farms_pan <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "PAN") %>%
  mutate(data_quality = 2)

write_csv(farms_pan, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_pan.csv"))


```


# Bangladesh
Current data is from ASC, only 1 farms, so definitely not comprehensive, just a drop in the bucket compared to the produciton of Bangladesh There are likely more farms than this. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

```{r}
farms_bgd <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "BGD") %>%
  mutate(data_quality = 2)

write_csv(farms_bgd, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_bgd.csv"))


```


# Guatemala 

Need to add 19 points to existing 3 (a final total of 22 points)

Current data is from the ASC, which is not comprehensive. 

http://www.fao.org/fishery/countrysector/naso_guatemala/en
 - “There are 34 shrimp farms of which 22 are active and 12 inactive.”
 
```{r}
farms_gtm <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "GTM") %>%
  mutate(data_quality = 4)

write_csv(farms_gtm, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_gtm.csv"))

## Now we will make a file that tells us the number of additional farms that need to be added to be comprehensive
gtm_additional <- data.frame(iso3c = "GTM", type = "shrimp", num_additional_farms = 19, data_quality = 4, total_farms = 22)
write.csv(gtm_additional, here("marine/shrimp/shrimp_farms/data/subset_known_add_points_4/add_farms_gtm.csv"), row.names = FALSE)
```


# Nicaragua

Seafood watch claims there is no available information other than the ASC farms. 
https://seafood.ocean.org/wp-content/uploads/2018/12/MBA_SeafoodWatch_FarmedShrimp_Nicaragua_Report.pdf

Current data is from ASC, only 2 farms, so definitely not comprehensive, just a drop in the bucket compared to the produciton of Nicaragua. There are likely more farms than this. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

```{r}
farms_nic <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "NIC") %>%
  mutate(data_quality = 2)

write_csv(farms_nic, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_nic.csv"))


```

# Australia
Current data is from ASC, only 2 farms, so probably not comprehensive. There are likely more farms than this. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

```{r}
farms_aus <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "AUS") %>%
  mutate(data_quality = 2)

write_csv(farms_aus, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_aus.csv"))


```


# Peru 

Need to add 71 points to existing 1 (a final total of 72 points).

Current data is a subset of farms from ASC. There is currently 1 point. 

https://seafood-tip.com/sourcing-intelligence/countries/peru/shrimp/
 - “22 Medium and large corporate farms are responsible for the majority of shrimp exports, while 50 small-scale farmers mainly produce for the domestic market”

```{r}
farms_per <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "PER") %>%
  mutate(data_quality = 4)

write_csv(farms_per, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_per.csv"))

## Now we will make a file that tells us the number of additional farms that need to be added to be comprehensive
per_additional <- data.frame(iso3c = "PER", type = "shrimp", num_additional_farms = 71, data_quality = 4, total_farms = 72)
write.csv(per_additional, here("marine/shrimp/shrimp_farms/data/subset_known_add_points_4/add_farms_per.csv"), row.names = FALSE)
```


# Colombia 
Need to add 19 points to existing 1 (a final total of 20 points)

Current data is from the ASC, which is a subset of points. 

http://www.fao.org/fishery/countrysector/naso_colombia/en
 - “The shrimp farming industry operates nine maturation laboratories, eighteen larval culture laboratories, **twenty shrimp farms** and seven processing plants. All of them employ qualified labour force, made up of professionals and expert technicians in their respective field of work (Minagricultura and ACUANAL, 2002).”
 
```{r}
farms_col <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "COL") %>%
  mutate(data_quality = 4)

write_csv(farms_col, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_col.csv"))

## Now we will make a file that tells us the number of additional farms that need to be added to be comprehensive
col_additional <- data.frame(iso3c = "COL", type = "shrimp", num_additional_farms = 19, data_quality = 4, total_farms = 20)
write.csv(col_additional, here("marine/shrimp/shrimp_farms/data/subset_known_add_points_4/add_farms_col.csv"), row.names = FALSE)
```


# Belize
 - There are 3 shrimp farms represented in the ASC data. However, the FAO report on aquaculture reports a number of other farms. Specifically: 
 - http://www.fao.org/fishery/countrysector/naso_belize/en
 -  “The fourteen shrimp farms operational are all located on the lower coastal plain of mainland Belize. The two northern-most of these farms (Nova Ladyville Shrimp Farm, Caribbean Shrimp Farm) are located in the Ladyville area, just north-east of the PSW Goldson International Airport. There are three shrimp farms (Crown Shrimp Limited, Paradise Shrimp Farm, Melinda Mariculture Limited) located between the Coastal Road and Dangriga area. There are two shrimp farms (Triton Mariculture, Haney's Shrimp Farm) situated between Dangriga and Riversdale. There are another four farms (Belize Aquaculture Limited, Royal Mayan Shrimp Farm, Texmar Shrimp Farm, Crustaceans Shrimp Farm) located between Riversdale and the lower reaches of the Placencia Lagoon. Three of these farms (Aquamar Shrimp Farm, Nova Toledo Shrimp Farm, Toledo Fish Farm) are also located in the Big Creek/Monkey River area.” 
 
 - Since we have the names of these… I will google and attempt to place on the map. The ASC already has 3 mapped (Aquamar shrimp farm, Cardelli farms, and Royal Mayan Shrimp Farms). Only 2 of these overlap, so we will assume that Cardelli has been added since the FAO report was published. 
 - I will need to place Nova Ladyville Shrimp Farm, Caribbean Shrimp Farm, Crown Shrimp Limited, Paradise Shrimp Farm, Melinda Mariculture Limited, Triton Mariculture, Haney's Shrimp Farm, Belize Aquaculture Limited, Texmar Shrimp Farm, Crustaceans Shrimp Farm, Nova Toledo Shrimp Farm, Toledo Fish Farm. If I can’t find these on google, then we will place the remaining number of farms randomly in suitable locations.
 
 
Nova Ladyville Shrimp Farm - This one has since been closed. 
Caribbean Shrimp Farm - 17.565826428687156, -88.30446591534489
Crown Shrimp Limited - Not on the map... I believe this has likely closed. 
Paradise Shrimp Farm - 17.088537866900786, -88.3141468760761
Melinda Mariculture Limited - 16.996401736714773, -88.28891758186664
Triton Mariculture - 16.91300830350531, -88.30401223336973
Haney's Shrimp Farm - 16.77084262777864, -88.3358879288162
Belize Aquaculture Limited - 16.664824857803115, -88.39502139821094
Texmar Shrimp Farm - 16.56115563573352, -88.41922498380852
Crustaceans Shrimp Farm - Not on the map... I believe this has likely closed. 
Nova Toledo Shrimp Farm - 16.460774477225453, -88.43760570009555
Toledo Fish Farm - 16.43434608053408, -88.4650001965993

```{r}

farms_blz <- data.frame(lat=c(17.565826428687156, 17.088537866900786, 16.996401736714773, 16.91300830350531, 16.77084262777864, 16.664824857803115, 16.56115563573352, 16.460774477225453, 16.43434608053408),
                        lon=c(-88.30446591534489, -88.3141468760761, -88.28891758186664, -88.30401223336973, -88.3358879288162, -88.39502139821094, -88.41922498380852, -88.43760570009555, -88.4650001965993),
                        subregion = c("Nova Ladyville Shrimp Farm", "Paradise Shrimp Farm", "Melinda Mariculture Limited", "Triton Mariculture", "Haney's Shrimp Farm", "Belize Aquaculture Limited", "Texmar Shrimp Farm", "Nova Toledo Shrimp Farm", "Toledo Fish Farm")) %>%
  mutate(iso3c = "BLZ") %>%
  mutate(data_quality = 5)

# now join with the ASC data
farms_blz_asc <- read.csv(here("marine/shrimp/shrimp_farms/data/ASC_shrimp_farm_lat_lon.csv")) %>%
  dplyr::filter(iso3c == "BLZ") 

farms_blz_total <- full_join(farms_blz, farms_blz_asc) 


write_csv(farms_blz_total, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_blz.csv"))

## visualize it 
farms_blz <-  st_as_sf(farms_blz_total, coords = c("lon", "lat"), 
                 crs = crs(gadm_latlon), agr = "constant")

mapview(farms_blz)
```

 
# Brazil
 - Has a map of major shrimp farming locations (from 2002): http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.200.2425&rep=rep1&type=pdf
 - Also has a table of area per state.
 
Source:
Mole, P., Bunge, J., 2002. Shrimp Farming in Brazil: An Industry Overview.

 - Because this source doesn't have specifics on numbers of farms or where they are located, I will treat this as no information that will need to be fully gapfilled. 


```{r}
bra_gf <- data.frame(iso3c = "BRA", type = "shrimp", num_additional_farms = NA, data_quality = 1)
write_csv(bra_gf, here("marine/shrimp/shrimp_farms/data/no_info_1/gf_bra.csv")) 
```
 
 
# Brunei
Note: These are all ponds on land, that use seawater. 

 - shrimp farms from this company: http://www.golden-corp.com/facilities.php?view=shrimp-farm-and-processing
 - the above map matches well with this... https://www.slideshare.net/ridzaludin/brunei-darussalam
 
 - ARDC: 4.716865, 114.535353: 14 hectares : Has a production table here: https://www.aquaculturealliance.org/advocate/brunei-project-develops-technology-for-large-black-tiger-shrimp-production-part-5/ : Uses seawater
 - Semaun Aquaculture: 4.722253, 114.541318: [40 hectares](https://pghjismail.wixsite.com/semaunaquaculture) : produce an average total of 60 metric tons of shrimp yearly : uses seawater
 - Keramut Shrimp Culture Industrial Site: 4.742089, 114.574539 : [92 hectares](https://borneobulletin.com.bn/2020/03/minister-visits-shrimp-farm-site-2/) :  : uses seawater
 - Penyatang Shrimp Culture Industrial Site: 4.761099, 114.610558: 200 hectares (from the map): uses seawater
 - Semaun Marine hatchery: 4.781321, 114.607222: 0.4 hectares (measured on google maps)
 - TMM phase 1/2: 4.970718, 115.043816: 51.471511 hectares (measured on google maps)
 
 
Source: 
Golden Corporation Shrimp Farm and Processing Map, n.d. http://www.golden-corp.com/facilities.php?view=shrimp-farm-and-processing


 - Because this is just data from one shrimp company in the Brunei, we will assume this is not comprehensive, and we will need to gapfill the remaining tonnage. 



```{r}
farms_brn <- data.frame(lat=c(4.716865, 4.722253, 4.742089, 4.761099, 4.781321, 4.970718),
                        lon=c(114.535353, 114.541318, 114.574539, 114.610558, 114.607222, 115.043816),
                        area_ha = c(14, 40, 92, 0.4, 200, 51.471511)) %>%
  mutate(iso3c = "BRN") %>%
  mutate(data_quality = 2) 

write_csv(farms_brn, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_brn.csv"))

## visualize it 
farms_brn <-  st_as_sf(farms_brn, coords = c("lon", "lat"), 
                 crs = crs(gadm_latlon), agr = "constant")

mapview(farms_brn)
```

 
# Cambodia 
 - There is a table in here. But it only has area estimates: https://clarklabs.org/wp-content/uploads/2019/01/Aquaculture-and-Coastal-Habitats-Report-No1.docx.pdf



Source: 
Eastman, J.R., Crema, S.C., Sangermano, F., Cunningham, S., Xiao, X., Hu, P., Johnson, C., Arakwiye, B., Crone, N., 2015. A Baseline Mapping of Aquaculture and Coastal Habitats in Thailand, Cambodia and Vietnam 14.

 - Because this source doesn't hav especifics on numbers of farms or where they are located, I will treat this as no information that will need to be fully gapfilled. 


```{r}
khm_gf <- data.frame(iso3c = "KHM", type = "shrimp", num_additional_farms = NA, data_quality = 1)
write_csv(khm_gf, here("marine/shrimp/shrimp_farms/data/no_info_1/gf_khm.csv")) 
```
 


# Costa Rica
 - This says there are 90 shrimp producers in Costa Rica. all located in Puntarenas and Guanacaste provinces: http://www.fao.org/fishery/countrysector/naso_costarica/en
 - We will place 90 farms into suitable locations along Costa Rica's coast.

Source: 
National Aquaculture Sector Overview. Costa Rica, 2005. , National Aquaculture Sector Overview Fact Sheets.

```{r}
farms_cri <- data.frame(iso3c = "CRI", type = "shrimp", num_additional_farms = 90, data_quality = 3)


write_csv(farms_cri, here("marine/shrimp/shrimp_farms/data/known_number_farms_3/farms_cri.csv"))

```


# Cuba

http://www.fao.org/fishery/countrysector/naso_cuba/en
 - "Marine cultures are being developed experimentally in conjunction with shrimp culture under the guidance of the Fisheries Research Center of the MIP.

Up-to-date, investment development and activities have been directed towards two zones:
Experimental Spawning Center at Cabo Cruz (some 100 km from Manzanillo).
 - 19.841789, -77.737212

The on-growing area in Niquero (70 km from Manzanillo), where packing facilities on land are available." 
 - 20.047065, -77.593449

Source: 
National Aquaculture Sector Overview. Cuba, 2005. National Aquaculture Sector Overview Fact Sheets.


https://edepot.wur.nl/531775

“Production of market-size shrimp takes place in 5 state-owned farms that are located at Cuba’s southern coast (which is less prone to the impact of storms than the country’s northern coast)”

This means that we need to place 3 more farms in Cuba. 

Source: 
Van Der Heijden, P., 2020. Aquaculture in Cuba. Aquaculture Europe 45.

```{r}
farms_cub <- data.frame(lat=c(19.841789, 20.047065),
                        lon=c(-77.737212, -77.593449),
                        area_ha = c(NA, NA)) %>%
  mutate(iso3c = "CUB") %>%
  mutate(data_quality = 4) 

write_csv(farms_cub, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_cub.csv"))

## visualize it 
farms_cub <-  st_as_sf(farms_cub, coords = c("lon", "lat"), 
                 crs = crs(gadm_latlon), agr = "constant")

mapview(farms_cub)

## Now we will make a file that tells us the number of additional farms that need to be added to be comprehensive
cub_additional <- data.frame(iso3c = "CUB", type = "shrimp", num_additional_farms = 3, data_quality = 4, total_farms = 5)
write.csv(cub_additional, here("marine/shrimp/shrimp_farms/data/subset_known_add_points_4/add_farms_cub.csv"), row.names = FALSE)
```


# El Salvador 
http://www.fao.org/fishery/countrysector/naso_elsalvador/en

 “According to the aquaculture method and the scale of production, aquaculture production units are distributed as follows: 15 production units of tilapia cages, 26 reservoirs for household consumption, 50 small and medium scale units, 1 industrial scale unit; 2 ornamental fish production units, 23 units of fresh water prawns, and 39 marine shrimp farms.”
 
 - I must’ve missed this my first lit review, so we will place 39 marine shrimp farms in suitable locations along the El Salvador coast. 

 
 Source: 
National Aquaculture Sector Overview. El Salvador, 2005. National Aquaculture Sector Overview Fact Sheets. Food and Agriculture Organization of the United Nations.
 
```{r}
farms_slv <- data.frame(iso3c = "SLV", type = "shrimp", num_additional_farms = 39, data_quality = 3)


write_csv(farms_slv, here("marine/shrimp/shrimp_farms/data/known_number_farms_3/farms_slv.csv"))
```
 
 
 
# Iran 
 - Claims 205 farms in 2000: http://library.enaca.org/Shrimp/Case/Africa/FinalAfrica.pdf
 
NASO Maps 
 - This is better information, so I will use these points. 
http://www.fao.org/fishery/naso-maps/selected-aquaculture-sites/iran/en/

Source: 
NASO aquaculture maps collection , 2010. Food and Agriculture Organization of the United Nations. [Accessed January 17, 2020] http://www.fao.org/fishery/naso-maps/selected-aquaculture-sites/iran/en/

 - We will use the existing points for exact locations, and then gapfill the remaining to reach a total of 205 farms. 

```{r}
## read in thailand shp file
iran_file <- file.path(path, "FAO_NASO/NASO/Point/Iran.kml")

iran_aqua <- read_sf(iran_file)

unique(iran_aqua$description)

mapview(iran_aqua)

## We need to filter for just the appropriate species of shrimp aquaculture

iran_finfish <- iran_aqua %>%
  dplyr::filter(!(str_detect(description, "Freshwater|freshwater"))) %>% ## Filter out freshwater systems
  dplyr::filter(!is.na(description))

mapview(iran_finfish) ## these are all marine shrimp farms now 

irn_bivalve_mol <- st_transform(iran_finfish, crs = "+init=epsg:3035") # convert to equal area projection to find centroids. 
irn_pts <- st_centroid(irn_bivalve_mol)

irn_pts_latlon <- st_transform(irn_pts, crs = "+init=epsg:4326") 

mapview(irn_pts_latlon)

farms_irn <- data.frame(do.call(rbind, st_geometry(irn_pts_latlon))) 
names(farms_irn) <- c("lon", "lat")
farms_irn <- farms_irn %>%
  mutate(iso3c="IRN") %>%
  mutate(data_quality = 4) 


write_csv(farms_irn, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_irn.csv"))


## Now we will make a file that tells us the number of additional farms that need to be added to be comprehensive
irn_additional <- data.frame(iso3c = "IRN", type = "shrimp", num_additional_farms = 191, data_quality = 4, total_farms = 205)
write.csv(irn_additional, here("marine/shrimp/shrimp_farms/data/subset_known_add_points_4/add_farms_irn.csv"), row.names = FALSE)

```

 
# Japan
 - No info at all... needs to be fully gapfilled.
    
Source: 
NA
    
```{r}
jpn_gf <- data.frame(iso3c = "JPN", type = "shrimp", num_additional_farms = NA, data_quality = 1)
write_csv(jpn_gf, here("marine/shrimp/shrimp_farms/data/no_info_1/gf_jpn.csv")) 
```


# Korea, Republic of


https://www.fftc.org.tw/htmlarea_file/activities/20110719101541/13.pdf
 - "West coast: shrimp (Penaeus japonicus, P. chinensis) is mainly cultivated in Taean, ChungNam"
 
Source: 
NA 

 - Because of the lack of information, this will need to be fully gapfilled. 

```{r}

kor_gf <- data.frame(iso3c = "KOR", type = "shrimp", num_additional_farms = NA, data_quality = 1)
write_csv(kor_gf, here("marine/shrimp/shrimp_farms/data/no_info_1/gf_kor.csv")) 
```
 
# Malaysia
 - [Seafood watch](https://www.seafoodwatch.org/-/m/sfw/pdf/reports/s/mba_seafoodwatch_malaysian_farmed_marine_shrimp_report.pdf) claims you can find the names and locations of registered shrimp farms here, but I can only find the broad suitable areas: https://www.dof.gov.my/ 
 - Broad suitable areas https://www.dof.gov.my/index.php/pages/view/78. I'll use these and use the area as a production estimate. Saved onto aurora /home/shares/food-systems/Food_footprint/_raw_data/Aquaculture_locations/Malaysia/malaysia_shrimp_farms.csv.
 
Source: 
Aquaculture Industrial Zone Area for High Impact Projects (HIP), n.d. Department of Fisheries Malaysia. [Accessed November 12, 2020] https://www.dof.gov.my/index.php/pages/view/78

 - This is only a subset of total shrimp farms most likely, since this only represents industrial scale projects. The rest will need to be gapfilled if necessary. 
 
```{r}
malaysia_shrimp_loc <- read_csv("/home/shares/food-systems/Food_footprint/_raw_data/Aquaculture_locations/Malaysia/malaysia_shrimp_farms.csv") %>%
  clean_names()

mys_area <- c(malaysia_shrimp_loc$luas_hektar)
```

Farms: 
 - AIR HANGAT, LANGKAWI, KEDAH: 6.433288, 99.826952
 - TAMAN AKUAKULTUR SEBATU MELAKA: 2.108604, 102.463536
 - TAMAN AKUAKULTUR LKIM TELAGA AIR, SARAWAK: 1.676758, 110.210032
 - KG. MERCHONG, PEKAN, PAHANG: 3.018046544221823, 103.43838831493285
 - KG. PENGKALAN GELAP, SETIU, TERENGGANU: 5.674088, 102.721589
 - KG. PASIR PUTEH, MARANG, TERENGGANU: 4.984545, 103.328932
 - PENARIK, SETIU, TERENGGANU: 5.597252273526056, 102.82739417198427
 - TANJUNG MANIS, SARAWAK: 2.197050, 111.183470
 - RUNGKUP HILIR, PERAK: 3.888025, 100.704595
 - SG. NIPAH, SABAK BERNAM, SELANGOR: 3.663462, 100.975557
 - TANJUNG BATU, PEKAN, PAHANG: 3.203638, 103.446051
 
Now lets map the farms 
```{r}
farms_mys <- data.frame(lat=c(6.433288, 2.108604, 1.676758, 3.018046544221823, 5.674088, 4.984545, 5.597252273526056, 2.197050, 3.888025, 3.663462, 3.203638),
                        lon=c(99.826952, 102.463536, 110.210032, 103.43838831493285, 102.721589, 103.328932, 102.82739417198427, 111.183470, 100.704595, 100.975557, 103.446051),
                        area_ha =mys_area,
                        subregion = c("Ayer Hangat, Kedah", "Sebatu, Malacca", "Kuching, Sarawak", "Pekan, Pahang", "Setiu, Terengganu", "Marang, Terengganu", "Setiu, Terengganu", "Tanjung Manis, Sarawak", "Rungkup, Perak", "Sabak, Selangor", "Pekan, Pahang")) %>%
  mutate(iso3c = "MYS") %>%
  mutate(data_quality = 2) 

write_csv(farms_mys, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_mys.csv"))

## visualize it 
farms_mys <-  st_as_sf(farms_mys, coords = c("lon", "lat"), 
                 crs = crs(gadm_latlon), agr = "constant")

mapview(farms_mys)


```
 

 
# Myanmar
 - http://www.fao.org/3/ad497e/ad497e05.htm
 "The largest area of shrimp farming is found in Rakhine State (155,533 acres), followed by Ayeyarwaddy (33 373 acres), Yangon (7 394 acres), with smaller areas in Bago, Kayin, Mon and Thaninthary"


Source: 
Myanmar aquaculture and inland fisheries, 2003. Food and Agriculture Organization of the United Nations Regional Office For Asia and the Pacific.

 - Because there is no other information, we will completely gapfill this country. 

```{r}
mmr_gf <- data.frame(iso3c = "MMR", type = "shrimp", num_additional_farms = NA, data_quality = 1)
write_csv(mmr_gf, here("marine/shrimp/shrimp_farms/data/no_info_1/gf_mmr.csv")) 
```


# Phillipines

https://juniperpublishers.com/ofoaj/pdf/OFOAJ.MS.ID.555668.pdf
"There are 271 brackish water shrimp farms (3,617.8ha) registered at the Bureau of Fisheries and Aquatic Resources of the Philippines as of 2013. "

 - We will place 271 farms in suitable locations along the philipino coast. 
 
Source: 
Vergel, C., 2017. Current Trends in the Philippines’ Shrimp Aquaculture lndustry:A Booming Blue Economy in the Pacific. OFOAJ 5. https://doi.org/10.19080/OFOAJ.2017.05.555668

```{r}

farms_phl <- data.frame(iso3c = "PHL", type = "shrimp", num_additional_farms = 271, data_quality = 3)


write_csv(farms_phl, here("marine/shrimp/shrimp_farms/data/known_number_farms_3/farms_phl.csv"))
```


# Saudi Arabia
http://www.naqua.com.sa/about-us.php
"Its shrimp hatchery can produce 5.5 billion nauplii and 2.5 billion postlarvae a year, and its 4,177 hectares of shrimp ponds along 65 kilometers of coastline are capable of producing *35,000* metric tons of shrimp a year, between 120 and 140 tons a day.  From harvest to processing takes less than four hours, and its plant can process up to 140 metric tons of shrimp a day." 
 - National aquaculture group: 16 shrimp farms (in one location): 7189 hectares (from google maps): 20.287253, 40.030864

https://www.alasmak.com.sa/en/about-us-en/index.html
 - Saudi Fisheries Company: *3000* tonnes per year: 17.806544, 41.897791: 206 hectares (taken from google maps)

Sources: 
National Aquaculture Group, n.d. [Accessed 11.12.20]. http://www.naqua.com.sa/about-us.php

Saudi Fish Company, n.d. [Accessed November 12, 2020]. https://www.alasmak.com.sa/en/about-us-en/index.html

 - Because of the large production out of SAU, this is probably not the only two farms. The remaining tonnage will be gapfilled. Data quality = 2

```{r}
farms_sau <- data.frame(lat=c(20.287253, 17.806544),
                        lon=c(40.030864, 41.897791)) %>%
  mutate(iso3c = "SAU") %>%
  mutate(data_quality = 2)

write_csv(farms_sau, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_sau.csv"))

## visualize it 
farms_sau <-  st_as_sf(farms_sau, coords = c("lon", "lat"), 
                 crs = crs(gadm_latlon), agr = "constant")

mapview(farms_sau)
```



# Sri Lanka 
http://www.fao.org/fishery/countrysector/naso_sri-lanka/en#:~:text=The%20current%20shrimp%20aquaculture%20industry,of%20around%203%20000%20ha.
"The current shrimp aquaculture industry in Sri Lanka is concentrated in the north western coastal belt **(Puttalam district)** covering a farm area of more than 4 500 ha with 70 hatcheries, of the total farm area the ponds themselves occupy an area of around 3 000 ha.... On the eastern coast in Batticaloa District, where shrimp farming in Sri Lanka first began in the late 1970s but was abandoned due to civil unrest, operations have recommenced in recent years. Over 60 small farms with an average farm area of between 1–2 ha were in operation at the end of 2002, with a total pond area of 155 ha."

 - We will place 130 shrimp farms in suitable locations along the Sri Lanka coast. 

Source: 
National Aquaculture Sector Overview. Sri Lanka, 2005. National Aquaculture Sector Overview Fact Sheets. Food and Agriculture Organization of the United Nations.


```{r}
farms_lka <- data.frame(iso3c = "LKA", type = "shrimp", num_additional_farms = 130, data_quality = 3)


write_csv(farms_lka, here("marine/shrimp/shrimp_farms/data/known_number_farms_3/farms_lka.csv"))

```


# Taiwan
https://www.taipeitimes.com/News/taiwan/archives/2020/06/19/2003738488
 - There are definitely a lot of farms
 
 - It is unclear how many farms are in Taiwan. This will be fully gapfilled. Data quality = 1. 
 
Source: 
NA
 
 
```{r}
twn_gf <- data.frame(iso3c = "TWN", type = "shrimp", num_additional_farms = NA, data_quality = 1)
write_csv(twn_gf, here("marine/shrimp/shrimp_farms/data/no_info_1/gf_twn.csv")) 

```


# United States of America
 
https://www.seafoodwatch.org/-/m/sfw/pdf/reports/s/mba_seafoodwatch_usfarmedshrimpreport.pdf
 - "Most commercial shrimp aquaculture production in the US is located in Texas, which has seven commercial farms. There is less production and fewer shrimp farms in Alabama (3), Florida (3), Hawaii (2), Nevada (1), Michigan (1), Indiana (1), Iowa (1), Maryland (1), and Massachusetts (1)."
 - We will only map for Alabama, Texas, Hawaii, and Florida. The rest are inland/indoor farms. 
 
Texas: 
http://www.texasaquaculture.org/TxShrimpFms.html
 - They have texas locations listed here. 
    - Bowers Shrimp Farm: 28.709786, -96.209726: 159.4461 hectares of ponds 
    - St. Martins Seafood: 28.731171, -96.203732: 80 hectares of ponds
    - Austwell Aqua Farm: 28.369096, -96.813910: 40 hectares
    - Southern Star Inc: 26.323330, -97.452599: 445 hectares
    - Harlinger Shrimp Farms: 26.149413, -97.322755: 182 hectares

Source:
Texas Aquaculture Association, 2018. [Accessed November 18, 2020] http://www.texasaquaculture.org/TxShrimpFms.html

 
Alabama: 
 - 3 locations according to SFW. From what I can tell, one of them is inland. And can only find 1 other.... 
 - Coordinates taken from google maps: 
    - Graham fisheries inc: 30.403144, -88.254304
    
Source: 
NA

Hawaii: 
 - 2 locations according to SFW.. I can only find one
     - [Kauai Shrimp](https://www.kauaishrimp.com/en/about-us/): 21.957597, -159.669555: 104 hectares
     
Source: 
NA
     
Florida: https://www.shrimpnews.com/FreeReportsFolder/NewsReportsFolder/USAflFiveShrimpFarmsAndHatchery.html
 - 3 marine locations according to SFW
    - [Gulf American Shrimp](https://www.gulfamericanshrimp.com/about-the-farm): 29.859915, -85.287438: 22 hectares
    - American Mariculture LLC: 26.602097, -82.113475: 3 hectares
    - Shrimp Improvement Systems: 24.972378, -80.555590: 0.88 hectares
    
Source: 

 - I will consider this comprehensive based on my research. 

```{r}
farms_usa <- data.frame(lat=c(28.709786, 28.731171, 28.369096, 26.323330, 26.149413, 30.403144, 21.957597,  29.859915, 26.602097, 24.972378),
                        lon=c(-96.209726, -96.203732, -96.813910, -97.452599, -97.322755, -88.254304, -159.669555, -85.287438, -82.113475, -80.555590),
                        area_ha = c(159.4461, 80, 40, 445, 182, NA, 104, 22, 3, 0.88)) %>%
  mutate(iso3c = "USA") %>%
  mutate(data_quality = 5) 

write_csv(farms_usa, here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5/farms_usa.csv"))

## visualize it 
farms_usa <-  st_as_sf(farms_usa, coords = c("lon", "lat"), 
                 crs = crs(gadm_latlon), agr = "constant")

mapview(farms_usa)
```


# Join cell count dataframes

```{r}
farm_files <- list.files(here::here("marine/shrimp/shrimp_farms/data/known_farm_sites_2_4_5"), full=TRUE)
all_farms = plyr::ldply(farm_files, read_csv)

all_farms_sf <- st_as_sf(all_farms, coords = c("lon", "lat"), 
                         crs = crs(gadm_latlon), agr = "constant")

mapview(all_farms_sf)

all_farms_1 <- all_farms %>%
  dplyr::select(-data_quality)

write.csv(all_farms_1, here::here("marine/shrimp/shrimp_farms/data/global_shrimp_farm_lat_lon.csv"), row.names = FALSE)
write.csv(all_farms, here::here("marine/shrimp/shrimp_farms/data/global_shrimp_farm_lat_lon_data_quality.csv"), row.names = FALSE)

```
